Copied files from Streamlit_projects required for cloud deployment
That other repo is the original where this and other apps are developed and maintained. This repo is only a copy of essential files for deployment to the Streamlit Community Cloud. This is because multiple apps can't be deployed from a single repo having unique config.toml files, so they need to be in separate repos.